requirements.txt
Django==5.1.2
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pillow==11.0.0
ipython==8.29.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
redis==5.2.0
stripe==11.2.0